+++ R/add_data.R
@@ -5,9 +5,9 @@
 #' @param study A study list object with class scivrs_study
 #' @param id The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it
 #' @param data The dataset as a data.frame, codebook object, path to a data file, or path to a codebook file
-#' @param vardesc Optional variable properties (see `codebook`)
+#' @param vardesc Optional variable properties (see `faux::codebook`)
 #' @param design A faux design specification
-#' @param ... Further dataset properties (see `codebook`)
+#' @param ... Further dataset properties (see `faux::codebook`)
 #'
 #' @return A study object with class scivrs_study
 #' @examples
@@ -17,7 +17,8 @@
 #'
 #' @export
 #'
-add_data <- function(study, id, data = NULL, vardesc = NULL, design = NULL, ...) {
+add_data <- function(study, id, data = NULL,
+                     vardesc = list(), design = NULL, ...) {
   id <- fix_id(id)
   d <- list(id = id)
 
@@ -44,7 +45,7 @@ add_data <- function(study, id, data = NULL, vardesc = NULL, design = NULL, ...)
   vm <- list()
   if (is.data.frame(data)) {
     # create codebook
-    cb <- codebook(data = data, name = id,
+    cb <- faux::codebook(data = data, name = id,
                    vardesc = vardesc, ...,
                    as_json = FALSE)
     d$codebook <- cb